Report collisions to Entity instances all at once.
Previously, collisions were reported to Entity instances one by one via
Entity.onCollision().  However, this did not allow the Entity to consider
all collisions at once and prioritize them if necessary.

To facilitate this, objects describing each collision are added to an
array.  The array is then passed to the Entity.handleCollisions() method.

Currently, the Entity.handleCollisions() method delegates each item in the
passed array to Entity.onCollision() and as such, some clean-up may be in
order.

Signed-off-by: suicidepills <atom.mac@gmail.com>